Get the table of individuals/nodes/edges/mutations from the tree sequence

This function extracts data from a given tree sequence table. All times are converted to model-specific time units from tskit's "generations backwards" time direction.

ts_table(ts, table = c("individuals", "edges", "nodes", "mutations"))

Arguments

ts

Tree sequence object of the class slendr_ts

table

Which tree sequence table to return

Value

Data frame with the information from the give tree-sequence table (can be either a table of individuals, edges, nodes, or mutations).

Details

For further processing and analyses, the output of the function ts_nodes might be more useful, as it merges the information in node and individual tables into one table and further annotates it with useful information from the model configuration data.

See also

Examples

check_dependencies(python = TRUE) # make sure dependencies are present

# load an example model with an already simulated tree sequence
model <- read_example("introgression")

# load the tree-sequence object from disk
ts <- ts_load(model, simplify = TRUE, mutate = TRUE, mutation_rate = 1e-8, random_seed = 42)

# get the 'raw' tskit table of individuals
ts_table(ts, "individuals")
#> # A tibble: 87 × 11
#>    ind_id  time pedigree_id raster_x raster_y pop_id alive remembered retained
#>     <dbl> <dbl>       <dbl>    <dbl>    <dbl>  <int> <lgl> <lgl>      <lgl>   
#>  1      0   -10    16023340        0        0      0 TRUE  TRUE       TRUE    
#>  2      1   -10    16023350        0        0      1 TRUE  TRUE       TRUE    
#>  3      2   -10    16023351        0        0      1 TRUE  TRUE       TRUE    
#>  4      3   -10    16023355        0        0      1 TRUE  TRUE       TRUE    
#>  5      4   -10    16023358        0        0      1 TRUE  TRUE       TRUE    
#>  6      5   -10    16023359        0        0      1 TRUE  TRUE       TRUE    
#>  7      6   -10    16023950        0        0      3 TRUE  TRUE       TRUE    
#>  8      7   -10    16025023        0        0      3 TRUE  TRUE       TRUE    
#>  9      8   -10    16025970        0        0      3 TRUE  TRUE       TRUE    
#> 10      9   -10    16027585        0        0      3 TRUE  TRUE       TRUE    
#> # … with 77 more rows, and 2 more variables: sampled <lgl>, time_tskit <dbl>

# get the 'raw' tskit table of edges
ts_table(ts, "edges")
#> # A tibble: 369 × 5
#>       id child parent  left  right
#>    <dbl> <int>  <int> <dbl>  <dbl>
#>  1     0     7     26     0 500000
#>  2     1    13     26     0 500000
#>  3     2     8     27     0 500000
#>  4     3    12     27     0 500000
#>  5     4     6     28     0 500000
#>  6     5    10     28     0 500000
#>  7     6    11     29     0 500000
#>  8     7    28     29     0 500000
#>  9     8     4     30     0 500000
#> 10     9    26     30     0 500000
#> # … with 359 more rows

# get the 'raw' tskit table of nodes
ts_table(ts, "nodes")
#> # A tibble: 102 × 5
#>    node_id ind_id pop_id  time time_tskit
#>      <int>  <int>  <int> <dbl>      <dbl>
#>  1       0     28      2 70010       2334
#>  2       1     28      2 70010       2334
#>  3       2     59      2 40010       1334
#>  4       3     59      2 40010       1334
#>  5       4      1      1   -10          0
#>  6       5      1      1   -10          0
#>  7       6      2      1   -10          0
#>  8       7      2      1   -10          0
#>  9       8      3      1   -10          0
#> 10       9      3      1   -10          0
#> # … with 92 more rows

# get the 'raw' tskit table of mutations
ts_table(ts, "mutations")
#> # A tibble: 2,183 × 5
#>       id  site  node     time time_tskit
#>    <dbl> <int> <int>    <dbl>      <dbl>
#>  1     0     0   100 3626753.    120892.
#>  2     1     1    35 5248391.    174947.
#>  3     2     2   100 1767187.     58907.
#>  4     3     3    35 2675842.     89195.
#>  5     4     4    35 4908016.    163601.
#>  6     5     5    35  704462.     23482.
#>  7     6     6   100 2477388.     82580.
#>  8     7     7    95  251623.      8388.
#>  9     8     8   100 4356631.    145221.
#> 10     9     9   100  876810.     29227.
#> # … with 2,173 more rows