Setup a dedicated Python virtual environment for slendr

Setup a dedicated Python virtual environment for slendr

setup_env(env = NULL, quiet = FALSE)

Arguments

env
Either a name of a conda environment, or a path to a standard Python virtual environment (such as one created by python3 -m venv <path to an environment>) with necessary Python dependencies (msprime, tskit, pyslim, and pandas). If NULL (the default), the user will be offered to install and configure an entirely isolated Miniconda Python environment just for slendr. If this environment is already present, calling setup_env() without the env argument will simply activate this environment.

quiet
Should informative messages be printed to the console? Default is FALSE.