Collect Identity-by-Descent (IBD) segments

This function iterates over a tree sequence and returns IBD tracts between pairs of individuals or nodes

Usage

ts_ibd(
  ts,
  coordinates = FALSE,
  within = NULL,
  between = NULL,
  minimum_length = NULL,
  maximum_time = NULL,
  sf = TRUE
)

Arguments

ts

Tree sequence object of the class slendr_ts

coordinates

Should coordinates of all detected IBD tracts be reported? If FALSE (the default), only the total length of shared IBD segments and their numbers are reported. If TRUE, coordinates of each segment will be returned (but note that this can have a massive impact on memory usage). See details for more information.

within

A character vector with individual names or an integer vector with node IDs indicating a set of nodes within which to look for IBD segments.

between

A list of lists of character vectors with individual names or integer vectors with node IDs, indicating a set of nodes between which to look for shared IBD segments.

minimum_length

Minimum length of an IBD segment to return in results. This is useful for reducing the total amount of IBD returned.

maximum_time

Oldest MRCA of a node to be considered as an IBD ancestor to return that IBD segment in results. This is useful for reducing the total amount of IBD returned.

sf

If IBD segments in a spatial tree sequence are being analyzed, should the returned table be a spatial sf object? Default is TRUE.

Value

A data frame with IBD results (either coordinates of each IBD segment shared by a pair of nodes, or summary statistics about the total IBD sharing for that pair)

Details

Iternally, this function leverages the tskit TreeSequence method ibd_segments. However, note that the ts_ibd function always returns a data frame of IBD tracts, it does not provide an option to iterate over individual IBD segments as shown in the official tskit documentation at https://tskit.dev/tskit/docs/stable/ibd.html. In general, R handles heavy iteration poorly, and this function does not attempt to serve as a full wrapper to ibd_segments.

Examples

check_dependencies(python = TRUE) # make sure dependencies are present

init_env()
#> The interface to all required Python modules has been activated.

# load an example model with an already simulated tree sequence
slendr_ts <- system.file("extdata/models/introgression.trees", package = "slendr")
model <- read_model(path = system.file("extdata/models/introgression", package = "slendr"))

# load the tree-sequence object from disk
ts <- ts_load(slendr_ts, model, simplify = TRUE)

# find IBD segments between specified Neanderthals and Europeans
ts_ibd(
  ts,
  coordinates = TRUE,
  between = list(c("NEA_1", "NEA_2"), c("EUR_1", "EUR_2")),
  minimum_length = 40000
)
#> # A tibble: 64 × 13
#>    node1 node2 length  mrca node1_time node2_time tmrca   left  right name1
#>    <int> <int>  <dbl> <dbl>      <dbl>      <dbl> <dbl>  <dbl>  <dbl> <chr>
#>  1     0    16  57475    66      70000          0 20001  26266  83741 NEA_1
#>  2     0    16  43592    66      70000          0 20001  83741 127333 NEA_1
#>  3     0    16 151679    66      70000          0 20001 225237 376916 NEA_1
#>  4     0    16  44211    66      70000          0 20001 387094 431305 NEA_1
#>  5     0    17  83741    66      70000          0 20001      0  83741 NEA_1
#>  6     0    17  54672    66      70000          0 20001 225237 279909 NEA_1
#>  7     0    17  97007    66      70000          0 20001 279909 376916 NEA_1
#>  8     0    17  45806    66      70000          0 20001 376916 422722 NEA_1
#>  9     0    18  57475    66      70000          0 20001  26266  83741 NEA_1
#> 10     0    18  43592    66      70000          0 20001  83741 127333 NEA_1
#> # ℹ 54 more rows
#> # ℹ 3 more variables: name2 <chr>, pop1 <fct>, pop2 <fct>