Recapitate the tree sequence

Recapitate the tree sequence

ts_recapitate(
  ts,
  recombination_rate,
  Ne,
  spatial = TRUE,
  migration_matrix = NULL,
  random_seed = NULL
)

Arguments

ts

Tree sequence object loaded by ts_load

recombination_rate

A constant value of the recombination rate

Ne

Effective population size during the recapitation process

spatial

Should spatial information encoded in the tree sequence data be converted to spatial R data structures? If FALSE, pixel-based raster-dimensions will not be converted to the coordinate reference system implied by the model. If TRUE (default), reprojection of coordinates will be performed. If the model was non-spatial, the value of this parameter is disregarded.

migration_matrix

Migration matrix used for coalescence of ancient lineages (passed to ts_recapitate)

random_seed

Random seed passed to pyslim's recapitate method

Value

Tree sequence object of the class slendr_ts

See also

Examples

if (FALSE) {
ts_rec <- ts_recapitate(ts, recombination_rate = 1e-8, Ne = 10000, random_seed = 42)
}