Get the table of individuals/nodes/edges/mutations from the tree sequence

This function extracts data from a given tree sequence table. All times are converted to model-specific time units from tskit's "generations backwards" time direction.

ts_table(ts, table = c("individuals", "edges", "nodes", "mutations"))

Arguments

ts

Tree sequence object of the class slendr_ts

table

Which tree sequence table to return

Value

Data frame with the information from the give tree-sequence table (can be either a table of individuals, edges, nodes, or mutations).

Details

For further processing and analyses, the output of the function ts_nodes might be more useful, as it merges the information in node and individual tables into one table and further annotates it with useful information from the model configuration data.

See also

Examples

# load an example model with an already simulated tree sequence
slendr_ts <- system.file("extdata/models/introgression.trees", package = "slendr")
model <- read_model(path = system.file("extdata/models/introgression", package = "slendr"))

# load the tree-sequence object from disk
ts <- ts_load(slendr_ts, model, simplify = TRUE, mutate = TRUE,
              mutation_rate = 1e-8, random_seed = 42)

# get the 'raw' tskit table of individuals
ts_table(ts, "individuals")
#> # A tibble: 50 × 11
#>    ind_id  time pedigree_id raste…¹ raste…² pop_id alive remem…³ retai…⁴ sampled
#>     <dbl> <dbl>       <dbl>   <dbl>   <dbl>  <int> <lgl> <lgl>   <lgl>   <lgl>  
#>  1      0   -10     5520341       0       0      0 TRUE  TRUE    TRUE    TRUE   
#>  2      1   -10     5520350       0       0      1 TRUE  TRUE    TRUE    TRUE   
#>  3      2   -10     5520353       0       0      1 TRUE  TRUE    TRUE    TRUE   
#>  4      3   -10     5520355       0       0      1 TRUE  TRUE    TRUE    TRUE   
#>  5      4   -10     5520357       0       0      1 TRUE  TRUE    TRUE    TRUE   
#>  6      5   -10     5520359       0       0      1 TRUE  TRUE    TRUE    TRUE   
#>  7      6   -10     5520480       0       0      3 TRUE  TRUE    TRUE    TRUE   
#>  8      7   -10     5520520       0       0      3 TRUE  TRUE    TRUE    TRUE   
#>  9      8   -10     5520711       0       0      3 TRUE  TRUE    TRUE    TRUE   
#> 10      9   -10     5520797       0       0      3 TRUE  TRUE    TRUE    TRUE   
#> # … with 40 more rows, 1 more variable: time_tskit <dbl>, and abbreviated
#> #   variable names ¹​raster_x, ²​raster_y, ³​remembered, ⁴​retained
#> # ℹ Use `print(n = ...)` to see more rows, and `colnames()` to see all variable names

# get the 'raw' tskit table of edges
ts_table(ts, "edges")
#> # A tibble: 105 × 5
#>       id child parent  left  right
#>    <dbl> <int>  <int> <dbl>  <dbl>
#>  1     0    14     26     0 500000
#>  2     1    15     26     0 500000
#>  3     2    11     27     0 500000
#>  4     3    13     27     0 500000
#>  5     4     4     28     0 500000
#>  6     5     5     28     0 500000
#>  7     6     9     29     0 500000
#>  8     7    10     29     0 500000
#>  9     8     7     30     0 500000
#> 10     9     8     30     0 500000
#> # … with 95 more rows
#> # ℹ Use `print(n = ...)` to see more rows

# get the 'raw' tskit table of nodes
ts_table(ts, "nodes")
#> # A tibble: 63 × 5
#>    node_id ind_id pop_id  time time_tskit
#>      <int>  <int>  <int> <dbl>      <dbl>
#>  1       0     19      2 70010       2334
#>  2       1     19      2 70010       2334
#>  3       2     21      2 40010       1334
#>  4       3     21      2 40010       1334
#>  5       4      1      1   -10          0
#>  6       5      1      1   -10          0
#>  7       6      2      1   -10          0
#>  8       7      2      1   -10          0
#>  9       8      3      1   -10          0
#> 10       9      3      1   -10          0
#> # … with 53 more rows
#> # ℹ Use `print(n = ...)` to see more rows

# get the 'raw' tskit table of mutations
ts_table(ts, "mutations")
#> # A tibble: 2,111 × 5
#>       id  site  node     time time_tskit
#>    <dbl> <int> <int>    <dbl>      <dbl>
#>  1     0     0    56  220923.     7364. 
#>  2     1     1    56  267788.     8927. 
#>  3     2     2    61  757678.    25256. 
#>  4     3     3    26 3622404.   120747. 
#>  5     4     4    61  694550.    23152. 
#>  6     5     5    61 5179659.   172656. 
#>  7     6     6    59  575947.    19199. 
#>  8     7     7    61 2373772.    79126. 
#>  9     8     8    26 1158835.    38628. 
#> 10     9     9    22    1480.       49.7
#> # … with 2,101 more rows
#> # ℹ Use `print(n = ...)` to see more rows